Reorganize package structure
